[fix] fixed a bug with id
